+++ lib/common/object/signPostObjectPolicyV4.d.ts
@@ -1,0 +1,1 @@
+export declare function signPostObjectPolicyV4(this: any, policy: string | object, date: Date): string;